+++ app/build.gradle.kts
@@ -75,6 +75,7 @@ ktlint {
 dependencies {
     implementation(project(":domain"))
     implementation(project(":data"))
+    implementation(project(":usecases"))
     implementation(libs.core.ktx)
     implementation(libs.lifecycle.runtime.ktx)
     implementation(libs.lifecycle.viewmodel.ktx)
