+++ app/src/main/kotlin/com/architects/pokearch/core/framework/network/PokemonServerDataSource.kt
@@ -2,8 +2,8 @@ package com.architects.pokearch.core.framework.network
 
 import android.util.Log
 import arrow.core.Either
-import com.architects.pokearch.core.domain.model.Pokemon
-import com.architects.pokearch.core.domain.model.error.Failure
+import com.architects.pokearch.domain.model.Pokemon
+import com.architects.pokearch.domain.model.error.Failure
 import com.architects.pokearch.core.framework.network.mappers.toDomain
 import com.architects.pokearch.core.framework.network.service.CryService
 import com.architects.pokearch.core.framework.network.service.PokedexService
