+++ app/src/main/kotlin/com/architects/pokearch/ui/features/home/ui/HomeItem.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.text.intl.Locale
 import coil.compose.AsyncImagePainter
 import coil.compose.rememberAsyncImagePainter
 import com.architects.pokearch.R
-import com.architects.pokearch.core.domain.model.Pokemon
+import com.architects.pokearch.domain.model.Pokemon
 import com.architects.pokearch.ui.components.animations.shimmerEffect
 import com.architects.pokearch.ui.components.extensions.GetColorsBackground
 import com.architects.pokearch.ui.components.extensions.buildImageRequest
