--- domain/src/main/java/com/architects/pokearch/domain/model/error/Failure.kt
@@ -1,5 +1,0 @@
-package com.architects.pokearch.core.domain.model.error
-
-sealed class Failure {
-    data object UnknownError: Failure()
-}
