--- domain/src/main/java/com/architects/pokearch/domain/repository/MediaPlayerRepositoryContract.kt
@@ -1,5 +1,0 @@
-package com.architects.pokearch.core.domain.repository
-
-interface MediaPlayerRepositoryContract {
-    suspend fun playCry(url: String)
-}
