+++ domain/src/main/kotlin/com/architects/pokearch/domain/model/PokemonInfo.kt
