+++ domain/src/main/kotlin/com/architects/pokearch/domain/model/error/Failure.kt
@@ -1,0 +1,5 @@
+package com.architects.pokearch.domain.model.error
+
+sealed class Failure {
+    data object UnknownError: Failure()
+}
