+++ domain/src/main/kotlin/com/architects/pokearch/domain/repository/MediaPlayerRepositoryContract.kt
@@ -1,0 +1,5 @@
+package com.architects.pokearch.domain.repository
+
+interface MediaPlayerRepositoryContract {
+    suspend fun playCry(url: String)
+}
