+++ domain/src/main/kotlin/com/architects/pokearch/domain/repository/SensorRepositoryContract.kt
