+++ settings.gradle.kts
@@ -17,3 +17,4 @@ rootProject.name = "PokeArch"
 include(":app")
 include(":domain")
 include(":data")
+include(":usecases")
